componentize everything
