+++ app/components/Views/Snaps/SnapSettings/SnapSettings.tsx
@@ -25,6 +25,7 @@ import Routes from '../../../../constants/navigation/Routes';
 import { Snap } from '@metamask/snaps-utils';
 import { getNavigationOptionsTitle } from '../../../UI/Navbar';
 import { useNavigation } from '@react-navigation/native';
+import { SnapDetails } from '../components/SnapDetails';
 
 interface SnapSettingsProps {
   snap: Snap;
@@ -39,7 +40,6 @@ const SnapSettings = () => {
   const navigation = useNavigation();
 
   const { snap } = useParams<SnapSettingsProps>();
-  const [enabled, setEnabled] = useState<boolean>(snap.enabled);
 
   useEffect(() => {
     navigation.setOptions(
@@ -53,91 +53,14 @@ const SnapSettings = () => {
     );
   }, [colors, navigation, snap.manifest.proposedName]);
 
-  const enableSnap = useCallback(async () => {
-    const { SnapController } = Engine.context as any;
-    await SnapController.enableSnap(snap.id);
-  }, [snap.id]);
-
-  const stopSnap = useCallback(async () => {
-    const { SnapController } = Engine.context as any;
-    await SnapController.stopSnap(snap.id);
-  }, [snap.id]);
-
   const removeSnap = useCallback(async () => {
     const { SnapController } = Engine.context as any;
     await SnapController.removeSnap(snap.id);
   }, [snap.id]);
 
-  const handleOnValueChange = useCallback(
-    (newValue) => {
-      setEnabled(newValue);
-      if (newValue) {
-        enableSnap();
-      } else {
-        stopSnap();
-      }
-    },
-    [enableSnap, stopSnap],
-  );
-
-  const versionBadge = () => (
-    <View style={styles.versionBadgeContainer}>
-      <Text
-        variant={TextVariant.HeadingSMRegular}
-        color={TextColor.Default}
-        style={styles.versionBadgeItem}
-      >
-        {snap.version}
-      </Text>
-      <Icon
-        name={IconName.Export}
-        size={IconSize.Sm}
-        style={styles.versionBadgeItem}
-      />
-    </View>
-  );
-
   return (
     <View style={styles.snapSettingsContainer}>
-      <View style={styles.snapInfoContainer}>
-        <Cell
-          style={styles.snapCell}
-          variant={CellVariants.Display}
-          title={snap.manifest.proposedName}
-          secondaryText={snap.id}
-          avatarProps={{
-            variant: AvatarVariants.Icon,
-            name: IconName.Snaps,
-          }}
-        />
-        <View style={styles.detailsContainerWithBorder}>
-          <Text variant={TextVariant.HeadingSM}>Enabled</Text>
-          <Switch
-            value={enabled}
-            trackColor={{
-              true: colors.primary.default,
-              false: colors.border.muted,
-            }}
-            onValueChange={handleOnValueChange}
-            ios_backgroundColor={colors.border.muted}
-          />
-        </View>
-        <View style={styles.detailsContainer}>
-          <Text variant={TextVariant.HeadingSM}>Install Origin</Text>
-          <View>
-            <Text variant={TextVariant.BodyMD} color={TextColor.Primary}>
-              {snap.versionHistory[0].origin}
-            </Text>
-            <Text variant={TextVariant.BodyMD} color={TextColor.Muted}>
-              {snap.versionHistory[0].date}
-            </Text>
-          </View>
-        </View>
-        <View style={styles.detailsContainer}>
-          <Text variant={TextVariant.HeadingSM}>Version</Text>
-          {versionBadge()}
-        </View>
-      </View>
+      <SnapDetails snap={snap} />
     </View>
   );
 };
