+++ app/components/Views/Snaps/components/SnapDetails/index.ts
@@ -1,0 +1,4 @@
+/* eslint-disable import/prefer-default-export */
+import SnapDetails from './SnapDetails';
+
+export { SnapDetails };
