+++ app/components/Views/Snaps/components/SnapVersionBadge/index.ts
@@ -1,0 +1,4 @@
+/* eslint-disable import/prefer-default-export */
+import SnapVersionBadge from './SnapVersionBadge';
+
+export { SnapVersionBadge };
